Define inversion in GF(2^32) using the algorithm schema of Itoh-Tsujii.
For this particular field extension, this algorithm is stunningly simple.
The main task of computing a^(r-1) turns out to be a nearly trivial computation,
requiring no multiplications in the base field at all!

Moreover, by extending the GF(2^16) table by a few more entries, we can
trim out some additional table lookups.  This is because we get lucky
with the irreducible polynomial for GF(2^32) because it's constant
term has logarithm 0x000d, a rather small constant.  We can extend
the exponential table enough to make sure that it suffices to
look up any (a + b + 0x000d) for a,b in [0..Q].  This makes the modular
reduction step computable with three table lookups instead of five.